Harden geolocation reserved IP handling
